Split fn_ctxt/adjust_fulfillment_errors from fn_ctxt/checks
